% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIpermSingleStepCT.R
\name{ARIpermCT}
\alias{ARIpermCT}
\title{ARI Permutation-based}
\usage{
ARIpermCT(copes, thr, mask, alpha, clusters,summary_stat,silent, family, delta, B, ct)
}
\arguments{
\item{copes}{list of copes map}

\item{thr}{threshold to construct cluster map}

\item{mask}{mask map, niftiImage class object or path}

\item{alpha}{alpha level}

\item{clusters}{clusters map as niftiImage class object or path, if NULL it is computed considering the threshold 3.2}

\item{summary_stat}{Choose among "max", "center-of-mass"}

\item{silent}{FALSE by default.}

\item{family}{which family for the confidence envelope? simes, finner, beta or higher.criticism. default is simes}

\item{delta}{do you want to consider at least delta size set?}

\item{B}{number of permutation, default 1000}

\item{ct}{set of thresholds}
}
\value{
Returns a list with the following objects: discoveries number of discoveries in the set selected, cluster id, maximum test statistic and relative coordinates
}
\description{
Performs ARI using permutation local test
}
\author{
Angela Andreella
}
