\name{ARIpermutation-package}
\alias{ARIpermutation-package}
\alias{ARIpermutation}
\docType{package}
\title{\packageTitle{ARIpermutation}}
\description{\packageDescription{ARIpermutation}}
\section{Package Content}{\packageIndices{ARIpermutation}}
\author{\packageAuthor{ARIpermutation}}
\section{Maintainer}{\packageMaintainer{ARIpermutation}}
\keyword{package}
