% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_threshold.R
\name{cluster_threshold}
\alias{cluster_threshold}
\title{Compute cluster map}
\usage{
cluster_threshold(map, max_dist)
}
\arguments{
\item{map}{Statistic Parametric Mapping used to create the cluster map}

\item{max_dist}{maximum distance to cut the tree}
}
\value{
Returns cluster map
}
\description{
given a threshold, e.g. 3.2, the function returns the map of clusters
}
\author{
Livio Finos
}
