% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalVector.R
\name{criticalVector}
\alias{criticalVector}
\title{Critical value}
\usage{
criticalVector(pvalues, family, alpha, lambda, delta = NULL)
}
\arguments{
\item{pvalues}{pvalues matrix with dimensiona variables times permutations}

\item{family}{family}

\item{alpha}{alpha}

\item{lambda}{lambda}

\item{delta}{delta}
}
\value{
critical value curve
}
\description{
compute critical values curve
}
\author{
Angela Andreella
}
