% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Critical value}
\usage{
cv(pvalues, family, alpha, lambda, ct = c(0,1), delta = NULL)
}
\arguments{
\item{pvalues}{pvalues raw}

\item{family}{family}

\item{alpha}{alpha}

\item{lambda}{lambda}

\item{ct}{set threshold}

\item{delta}{delta}
}
\value{
critical value curve
}
\description{
compute critical values curve
}
\author{
Angela Andreella
}
