% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dI.R
\name{dI}
\alias{dI}
\title{Lower bound true discoveries}
\usage{
dI(ix, cv, praw)
}
\arguments{
\item{ix}{= set-wise of hypotheses considered}

\item{cv}{= vector of critical values}

\item{praw}{= raw pvalues}
}
\value{
Returns the lower confidence bound
}
\description{
Calculates (1-alpha) lower confidence bounds for the set-wise of false null hypotheses
}
\author{
Angela Andreella
}
