% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoveriesPerm.R
\name{discoveriesPerm}
\alias{discoveriesPerm}
\title{Lower bound for the number of true discoveries}
\usage{
discoveriesPerm(out, ix)
}
\arguments{
\item{out}{output from the \code{\link{pARI}} function.}

\item{ix}{numeric vector. It refers to the set-wise hypotheses considered.}
}
\value{
discoveriesPerm
}
\description{
Internal function.
}
\author{
Angela Andreella
}
