% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIpermSingleStepCT.R, R/cv.R, R/lambdaOpt.R,
%   R/plotNullDistribution.R
\docType{data}
\name{family_set}
\alias{family_set}
\title{ARI Permutation-based}
\format{An object of class \code{character} of length 4.}
\usage{
family_set

family_set

family_set

family_set
}
\arguments{
\item{thr}{= threshold to construct cluster map}

\item{mask}{= mask map, niftiImage class object or path}

\item{clusters}{= clusters map as niftiImage class object or path, if NULL it is computed considering the threshold 3.2}

\item{summary_stat}{Choose among \code{=c("max", "center-of-mass")}}

\item{silent}{\code{FALSE} by default.}

\item{B}{number of permutation, default 1000}

\item{lambda}{lambda}

\item{pvalues}{pvalues raw}

\item{P}{permutation matrix pvalues where rows correspond to the permutation, default NULL}

\item{family}{which family for the confidence envelope? simes, finner, beta or higher.criticism. default is NULL}

\item{alpha}{type I error allowed to construct the critical family, default 0.1}

\item{ct}{set of threshold to construct the critical family, default \code{c(0,1)}}

\item{path}{selected the path where the plot will be saved, if NULL the current working directory is used}

\item{name}{name plot pdf, if NULL plot is used}

\item{delta}{for the family critical values, default NULL}

\item{copes}{image copes instead of pvalues, default NULL}
}
\value{
Returns a list with the following objects: \code{discoveries} number of discoveries in the set selected, cluster id, maximum test statistic and relative coordinates

critical value curve

lambda

Returns plot null distribution with critical value curve and observed pvalues in red
}
\description{
Performs ARI using permutation local test

compute critical values curve

compute lambda parameter
}
\author{
Angela Andreella

Angela Andreella

Angela Andreella
}
\keyword{datasets}
