% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_array.R
\name{get_array}
\alias{get_array}
\title{get_array}
\usage{
get_array(map,map_dims=NULL)
}
\arguments{
\item{map}{array, character giving the nii file (address and) name.}

\item{map_dims}{vector with the 3 dimension the map must agree. if \code{NULL} (default) no checks are made}
}
\value{
array
}
\description{
get the array from a map parameter. and make compatibility checks
}
\author{
Angela Andreella
}
