% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histP.R
\name{histP}
\alias{histP}
\title{Histogram Pvalues}
\usage{
histP(copes, alternative, mask, zstat, path, name, method, rand, B)
}
\arguments{
\item{copes}{image copes instead of pvalues, default NULL}

\item{alternative}{alternative, default is "two.sided"}

\item{mask}{mask, default is NULL}

\item{zstat}{zstat, default is NULL}

\item{path}{path}

\item{name}{name}

\item{method}{method to compute the zstat}

\item{rand}{logical. Should p values computed by permutation distribution?}

\item{B}{number of permutations to perform, default is 1000, if matrix p-value is computed directly.}
}
\value{
Returns hist
}
\description{
create histogram p-values
}
