% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaPlot.R
\name{lambdaPlot}
\alias{lambdaPlot}
\title{Lambda plot}
\usage{
lambdaPlot(copes,family,ct,alpha, delta,P, mask, path, name)
}
\arguments{
\item{copes}{list of copes, default NULL}

\item{family}{family}

\item{ct}{set threshold}

\item{alpha}{alpha}

\item{delta}{delta}

\item{P}{matrix of pvalues where rows indicate the permutations, default NULL}

\item{mask}{mask}

\item{path}{path}

\item{name}{name}
}
\value{
lambda
}
\description{
compute lambda calibration plot
}
\author{
Angela Andreella
}
