% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_TDP.R
\name{map_TDP}
\alias{map_TDP}
\title{True Discovery Proportion brain map}
\usage{
map_TDP(ARIout,path,name,mask)
}
\arguments{
\item{ARIout}{Output object by \code{\link{pARIbrain}}.}

\item{path}{Character string. Path to save the NIfTI file. The path does not must end with \code{/}.}

\item{name}{Character string. The name of the map NIfTI file that will be used.}

\item{mask}{NIfTI file or character string. 3D array of logical values (i.e. \code{TRUE/FALSE} in/out of the brain). 
Alternatively it may be a (character) NIfTI file name. If \code{mask=NULL}, it is assumed that none of the voxels have to be excluded.}
}
\value{
The function write directly in the \code{path} specified the true discovery proportion NIfTI map with name specified in \code{name}.
}
\description{
Performs the True Discovery Proportion brain map.
}
\author{
Angela Andreella
}
