% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneSamplePar.R
\name{oneSamplePar}
\alias{oneSamplePar}
\title{One-Sample t-test}
\usage{
oneSamplePar(X,alternative= "two.sided")
}
\arguments{
\item{X}{data matrix where rows represent the \code{m} variables and columns the \code{n} observations.}

\item{alternative}{character string. It refers to the alternative hypothesis, must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"lower"}.}
}
\value{
Returns a list with the following objects: 
- \code{Test}: vector with length equals \eqn{m}. Observed two-sample t-tests, one for each \eqn{m} variable, 
- \code{pv}: vector with length equals \eqn{m}. observed p-values, one for each \eqn{m} variable,
}
\description{
Performs parametric one-sample t-tests.
}
\author{
Angela Andreella
}
