% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pARIgene.R
\name{pARIgene}
\alias{pARIgene}
\title{Permutation-based All-Resolutions Inference for Gene Expression Data}
\usage{
pARIgene(X= NULL, pathways, alpha = 0.05, family = "simes", delta = 0, 
B = 1000, test.type = "one_sample", complete = FALSE, iterative = FALSE, 
approx = TRUE, ncomb = 100, step.down = FALSE, max.step = 10, ...)
}
\arguments{
\item{X}{data matrix where rows represent the \eqn{m} variables and columns 
the \eqn{n} observations.}

\item{pathways}{list of pathways where names indicates the name of the pathway.}

\item{alpha}{numeric value in `[0,1]`. It expresses the alpha level to control 
the family-wise error rate.}

\item{family}{string character. Choose a family of confidence envelopes to 
compute the critical vector from \code{"simes"}, \code{"aorc"}, \code{"beta"} 
and \code{"higher.criticism"}.#' @param alpha alpha level.}

\item{delta}{numeric value. It expresses the delta value, 
please see the references. Default to 0.}

\item{B}{numeric value. Number of permutations, default to 1000.}

\item{test.type}{character string. Choose a type of tests among 
\code{"one_sample"}, i.e., one-sample t-test, or \code{"two_samples"}, 
i.e., two-samples t-tests. Default \code{"one_sample"}.}

\item{complete}{Boolean value. If \code{TRUE} the sets of critical vectors 
and the raw pvalues are returned. Default @FALSE.}

\item{iterative}{Boolean value. If \code{iterative = TRUE}, the iterative 
method for improvement of confidence envelopes is applied. Default @FALSE.}

\item{approx}{Boolean value. Default @TRUE. If you are treating high 
dimensional data, we suggest to put \code{approx = TRUE} to speed up 
the computation time.}

\item{ncomb}{Numeric value. If \code{approx = TRUE}, you must decide 
how many random subcollections (level of approximation) considered.}

\item{step.down}{Boolean value. Default @FALSE If you want to compute 
the lambda calibration parameter using the step-down approach put \code{TRUE}.}

\item{max.step}{Numeric value. Default to 10. Maximum number of steps for 
the step down approach, so useful when \code{step.down = TRUE}.}

\item{...}{Futher parameters.}
}
\value{
by default returns a list with the following objects: 
\code{discoveries}: lower bound for the number of true discoveries 
in the set selected, \code{ix}: selected variables. I
f \code{complete = TRUE} the raw \code{pvalues} and \code{cv} 
critical vector are returned.
}
\description{
This function computes the lower bound for the  number of true 
discoveries within each cluster (pathways) of Gene Expression Data.
}
\references{
For the general framework of All-Resolutions Inference see:

Goeman, Jelle J., and Aldo Solari. "Multiple testing for exploratory research.
" Statistical Science 26.4 (2011): 584-597.

For permutation-based All-Resolutions Inference see:

Andreella, Angela, et al. "Permutation-based true discovery proportions for 
fMRI cluster analysis." arXiv preprint arXiv:2012.00368 (2020).
}
\seealso{
The type of tests implemented: \code{\link{signTest}} 
\code{\link{permTest}}.
}
\author{
Angela Andreella
}
