% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rejPerm}
\alias{rejPerm}
\title{n rejections using permutation theory}
\usage{
rejPerm(pvalues,ct)
}
\arguments{
\item{pvalues}{matrix with dimension number of permutation times number of tests}

\item{ct}{set of cutoff between 0 and 1}
}
\value{
Returns number of rejections
}
\description{
Compute the number of rejections having pvalues computed using permutation distribution
}
\author{
Angela Andreella
}
