% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowVariance.R
\name{rowVariance}
\alias{rowVariance}
\title{Rows Variance}
\usage{
rowVariance(X,na.rm = TRUE)
}
\arguments{
\item{X}{data matrix where rows represents the variables and columns the observations.}

\item{na.rm}{by default \code{na.rm=TRUE}.}
}
\value{
rows variance
}
\description{
performs the variance for each row in a matrix.
}
\author{
Angela Andreella
}
