% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signTest.R
\name{signTest}
\alias{signTest}
\title{Sign-Flipping Test}
\usage{
signTest(X, B = 1000, alternative = "two.sided", seed = NULL, mask = NULL, rand = F)
}
\arguments{
\item{X}{data where rows represents the variables and columns the observations}

\item{B}{number of permutations to perform, default is 1000.}

\item{alternative}{character referring to the alternative hypothesis, "two.sided", "greater" or "lower". Default is "two.sided"}

\item{seed}{specify seed, default is 1234.}

\item{mask}{mask to apply, it can be nii format or path}

\item{rand}{logical. Should p values computed by permutation distribution?}
}
\value{
Returns a list with the following objects: \code{Test} observed one sample t-test, \code{Test_H0} Test statistics under H0, \code{pv} observed p-values, \code{pv_H0} p-values under H0
}
\description{
Performs sign-flipping, i.e. permutation, one-sample t-tests
}
\author{
Angela Andreella
}
