% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCE.R
\name{simCE}
\alias{simCE}
\title{Simulation Confidence Envelope}
\usage{
simCE(m,n,pi0,B,rho,SNR,alpha,delta,ct,family)
}
\arguments{
\item{m}{Number of variables, i.e. hypotheses}

\item{n}{Number of observations}

\item{pi0}{Proportion of true null hypotheses}

\item{B}{number of permutations}

\item{rho}{Level of equi-correlation between pairs of variables}

\item{SNR}{Signal to noise ratio}

\item{alpha}{significance level}

\item{delta}{which size set do you want to consider at least?}

\item{ct}{set of threshold}

\item{family}{family of confidence bound}
}
\value{
simulation CE
}
\description{
Simulation Confidence Envelope using the gaussianSamples function from the package sansSouci
}
