% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{simulate normal distributed data}
\usage{
simulateData(pi0,m,rho,SNR, set.seed = NULL)
}
\arguments{
\item{pi0}{proportion of true null hypothesis}

\item{m}{number of variables, i.e., hypothesis}

\item{rho}{Level of equi-correlation between pairs of variables}

\item{SNR}{Signal to noise ratio}

\item{set.seed}{specify seed}
}
\value{
Returns a list with the following objects: \code{Test} observed one sample t-test, \code{Test_H0} Test statistics under H0, \code{pv} observed p-values, \code{pv_H0} p-values under H0
}
\description{
Simulate normal distributed data
}
\author{
Angela Andreella
}
