% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_perm_roi.R
\name{summary_perm_roi}
\alias{summary_perm_roi}
\title{Summary permutation ROI}
\usage{
summary_perm_roi(cv, ix, pvalues)
}
\arguments{
\item{cv}{= vector of critical values}

\item{ix}{= vector of index regarding the set of hypothesis to analyze}

\item{pvalues}{= raw pvalues}
}
\value{
Returns a list with the following objects: \code{Size} Total number of hypotheses rejected, \code{FalseNull} number of false null hypotheses, \code{TrueNull} number of true null hypotheses, and \code{ActiveProp} active proportion inside ix.
}
\description{
Performs Single step closed testing using permutation local test
}
\author{
Angela Andreella
}
