% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDPerm.R
\name{TDPerm}
\alias{TDPerm}
\title{True Discovery Permutation}
\usage{
TDPerm(ix, cv, p)
}
\arguments{
\item{ix}{The selection of hypotheses considered}

\item{cv}{critical values}

\item{p}{raw pvalues}
}
\value{
lower bounds discoveries
}
\description{
performs lower confidence bound of true discoveries using permutation local test.
}
\author{
Angela Andreella
}
