% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneSample.R, R/permTest.R, R/signTest.R
\docType{data}
\name{alternative_set}
\alias{alternative_set}
\title{One Sample t test}
\format{
An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.
}
\usage{
oneSample(X,alternative= "two.sided")

permTest(X, B, alternative, seed, mask, rand, label)

signTest(X, B = 1000, alternative = "two.sided", seed = NULL, mask = NULL, rand = F)
}
\arguments{
\item{label}{labels of the observations, if NULL the columns's name are considered, default NULL}

\item{X}{data where rows represents the variables and columns the observations}

\item{B}{number of permutations to perform, default is 1000.}

\item{alternative}{character referring to the alternative hypothesis, "two.sided", "greater" or "lower". Default is "two.sided"}

\item{seed}{specify seed, default is 1234.}

\item{mask}{mask to apply, it can be nii format or path}

\item{rand}{logical. Should p values computed by permutation distribution?}
}
\value{
Returns a list with the following objects: \code{Test} test statistic, and \code{pv} corresponding raw pvalues.

Returns a list with the following objects: \code{Test} observed one sample t-test, \code{Test_H0} Test statistics under H0, \code{pv} observed p-values, \code{pv_H0} p-values under H0

Returns a list with the following objects: \code{Test} observed one sample t-test, \code{Test_H0} Test statistics under H0, \code{pv} observed p-values, \code{pv_H0} p-values under H0
}
\description{
Performs One Sample t test

Performs permutation tests, i.e. permutation, two-sample t-tests

Performs sign-flipping, i.e. permutation, one-sample t-tests
}
\author{
Angela Andreella

Angela Andreella

Angela Andreella
}
\keyword{datasets}
