% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalVector.R
\name{criticalVector}
\alias{criticalVector}
\title{Critical value}
\usage{
criticalVector(pvalues, family, alpha, lambda, delta = NULL, m = NULL)
}
\arguments{
\item{pvalues}{pvalues matrix with dimensiona variables times permutations}

\item{family}{Choose a family of confidence envelopes to compute the critical vector from \code{"simes"}, \code{"aorc"}, \code{"beta"} and \code{"higher.criticism"}.}

\item{alpha}{alpha}

\item{lambda}{lambda}

\item{delta}{delta}

\item{m}{number of hypothesis}
}
\value{
critical value curve
}
\description{
compute critical values curve
}
\author{
Angela Andreella
}
