% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dI.R
\name{dI}
\alias{dI}
\title{Lower bound true discoveries}
\usage{
dI(ix,cv,praw)
}
\arguments{
\item{ix}{set-wise of hypotheses considered}

\item{cv}{vector of critical values}

\item{praw}{raw pvalues}

\item{iterative}{if \code{iterative = TRUE}, the iterative iterative method for improvement of confidence envelopes is applied.}

\item{approx}{if \code{iterative = TRUE} and you are treating high dimensional data, we suggest to put \code{approx = TRUE} to speed up the computation time.}

\item{ncomb}{if \code{approx = TRUE}, you must decide how many large random subcollection (level of approximation) considered.}
}
\value{
Returns the lower confidence bound
}
\description{
Calculates (1-alpha) lower confidence bounds for the set-wise of false null hypotheses
}
\author{
Angela Andreella
}
