% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoveriesPerm.R
\name{discoveriesPerm}
\alias{discoveriesPerm}
\title{discoveriesPerm}
\usage{
discoveriesPerm(out, ix)
}
\arguments{
\item{out}{out}

\item{ix}{ix}
}
\value{
discoveriesPerm
}
\description{
compute discoveriesPerm
}
\author{
Angela Andreella
}
