% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histP.R
\name{histP}
\alias{histP}
\title{Histogram Pvalues}
\usage{
histP(copes, alternative, mask, zstat, path, name)
}
\arguments{
\item{copes}{image copes instead of pvalues, default NULL}

\item{alternative}{alternative}

\item{mask}{mask}

\item{zstat}{zstat}

\item{path}{path}

\item{name}{name}

\item{method}{method to compute the zstat}
}
\value{
Returns hist
}
\description{
create histogram p-values
}
