% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaOpt.R
\name{lambdaOpt}
\alias{lambdaOpt}
\title{Lambda calibration}
\usage{
lambdaOpt(pvalues,family,alpha, delta)
}
\arguments{
\item{pvalues}{pvalues matrix with dimensions equal to the number of variables times the number of permutations.}

\item{family}{by default \code{family="simes"}. Choose a family of confidence envelopes to compute the critical vector from \code{"simes"}, \code{"AORC"}, \code{"beta"} and \code{"higher.criticism"}.}

\item{alpha}{alpha level.}

\item{delta}{by default \code{delta = 0}. Do you want to consider sets with at least delta size?}
}
\value{
lambda parameter
}
\description{
\code{lambdaOpt} provides the optimal lambda calibration parameter used in the critical vector.
}
\author{
Angela Andreella
}
