% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaOpt.R
\name{lambdaOpt}
\alias{lambdaOpt}
\title{Lambda calibration}
\usage{
lambdaOpt(pvalues,family,alpha, delta = 0)
}
\arguments{
\item{pvalues}{pvalues matrix with dimensions variables times permutations}

\item{family}{family}

\item{alpha}{alpha}

\item{delta}{delta}
}
\value{
lambda
}
\description{
compute lambda parameter
}
\author{
Angela Andreella
}
