% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_TDP.R
\name{map_TDP}
\alias{map_TDP}
\title{map True Discovery Proportion}
\usage{
map_TDP(ARIout, path = getwd(), name = "tdp", mask)
}
\arguments{
\item{ARIout}{= output object ARI permutation}

\item{path}{= path}

\item{name}{= name}

\item{mask}{= mask map, niftiImage class object or path}
}
\value{
Returns a TDP map
}
\description{
Performs brain map True Discovery Proportion
}
\author{
Angela Andreella
}
