% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneSample.R
\name{oneSample}
\alias{oneSample}
\title{One Sample t test}
\usage{
oneSample(X,alternative)
}
\arguments{
\item{X}{data where rows represents the variables and columns the observations}

\item{alternative}{= character referring to the alternative hypothesis, "two.sided", "greater" or "less". Default is "two.sided"}
}
\value{
Returns a list with the following objects: \code{Test} test statistic, and \code{pv} corresponding raw pvalues.
}
\description{
Performs One Sample t test
}
\author{
Angela Andreella
}
