% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pARI.R
\name{pARI}
\alias{pARI}
\title{Single Step permutation-based method}
\usage{
pARI(X,ix, alpha, family, delta, B, pvalues, test.type, complete, ...)
}
\arguments{
\item{X}{data matrix rows represent variables, columns observations, the first rows are the raw pvalues.}

\item{ix}{set of hypothesis of interest. It can be a vector having same lenght of the variables or the indices if one set is considered.}

\item{alpha}{alpha level, default 0.1}

\item{family}{family of confidence envelopes considered in order to find the critical values. Now, we implement the Simes ones and the Beta}

\item{delta}{do you want to consider at least delta size set?}

\item{B}{number of permutations}

\item{pvalues}{matrix pvalues instead of data with dimensions hypotheses times permutations, default NULL}

\item{test.type}{"one_sample", i.e., one sample, or "two_samples", i.e., two samples  t-tests?}

\item{complete}{default FALSE. If TRUE the sets of critical vectors and raw pvalues are returned.}

\item{...}{Futher arguments, see details.}
}
\value{
Returns a list with the following objects discoveries number of discoveries in the set selected, pvalues raw pvalues
}
\description{
Performs single step method based on confidence envelope constructed by sign flipping
}
\seealso{
\code{\link{signTest}} \code{\link{permTest}}
}
\author{
Angela Andreella
}
