% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pARIbrain.R
\name{pARIbrain}
\alias{pARIbrain}
\title{Permutation-based All-Resolutions Inference for brain imaging}
\usage{
pARIbrain(copes, thr, mask, alpha, clusters, 
alternative, summary_stat, silent, family, delta, B, rand)
}
\arguments{
\item{copes}{The list of copes, i.e., constrasts maps, one for each subject used to compute the statistical tests.}

\item{thr}{threshold used to construct the cluster map.}

\item{mask}{3D array of locicals (i.e. \code{TRUE/FALSE} in/out of the brain). 
Alternatively it may be a (character) NIfTI file name. If \code{mask=NULL}, it is assumed that non of the voxels have to be excluded.}

\item{alpha}{alpha level}

\item{clusters}{3D array of cluster ids (0 when voxel does not belong to any cluster) or a (character) nifti file name. 
If \code{cluster=NULL} the cluster map is computed by the \code{\link{cluster_threshold}} function with threshold equals 3.2.}

\item{alternative}{a character string referring to the alternative hypothesis, must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"lower"}.}

\item{summary_stat}{Choose among \code{=c("max", "center-of-mass")}.}

\item{silent}{\code{FALSE} by default.}

\item{family}{by default \code{family="simes"}. Choose a family of confidence envelopes to compute the critical vector from \code{"simes"}, \code{"finner"}, \code{"beta"} and \code{"higher.criticism"}.}

\item{delta}{by default \code{delta = 0}. Do you want to consider sets with at least delta size?}

\item{B}{by default \code{B = 1000}. Number of permutations.}

\item{rand}{by default \code{rand = FALSE}. 
A logical value, if \code{rand = TRUE}, the pvalues are computed by \code{\link{rowRanks}}.}
}
\value{
Returns a matrix with the following objects: 
Size, FalseNull, TrueNull, ActiveProp and other statistics for each cluster.
}
\description{
The main function for single step All-Resolutions Inference (ARI) method based on critical vectors constructed by permutations for fMRI cluster analysis.
}
\examples{
\dontrun{
library(fMRIdata)
data(Auditory_clusterTH3_2)
data(Auditory_copes)
data(Auditory_mask)
auditory_out <- pARIbrain(copes = Auditory_copes, 
cluster = Auditory_clusterTH3_2, mask = Auditory_mask, 
alpha = 0.05, silent = TRUE)
auditory_out$out
}
}
\references{
For the general framework of All-Resolutions Inference see:

Goeman, Jelle J., and Aldo Solari. "Multiple testing for exploratory research." Statistical Science 26.4 (2011): 584-597.

For All-Resolutions Inference for functional Magnetic Resonance Imaging data see: 

Rosenblatt, Jonathan D., et al. "All-resolutions inference for brain imaging." Neuroimage 181 (2018): 786-796.


For permutation-based All-Resolutions Inference see:

Andreella, Angela, et al. "Permutation-based true discovery proportions for fMRI cluster analysis." arXiv preprint arXiv:2012.00368 (2020).
}
\author{
Angela Andreella
}
