% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permTest.R
\name{permTest}
\alias{permTest}
\title{Permutation Test}
\usage{
permTest(X, B, alternative, seed, mask, rand, label)
}
\arguments{
\item{X}{data where rows represents the variables and columns the observations, the columns's name define the categories}

\item{B}{by default \code{B = 1000}. Number of permutations.}

\item{alternative}{a character string referring to the alternative hypothesis, must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"lower"}.}

\item{seed}{by default \code{seed=1234}. Integer value specifying the seed.}

\item{mask}{3D array of locicals (i.e. \code{TRUE/FALSE} in/out of the brain). Alternatively it may be a (character) NIfTI file name. If \code{mask=NULL}, it is assumed that non of the voxels have to be excluded.}

\item{rand}{by default \code{rand = FALSE}.}

\item{label}{by default \code{label = NULL}. Labels of the observations, if \code{NULL} the columns's name are considered.}
}
\value{
Returns a list with the following objects: \code{Test} observed one sample t-test, \code{Test_H0} Test statistics under H0, \code{pv} observed p-values, \code{pv_H0} p-values under H0
}
\description{
Performs permutation tests, i.e. permutation, two-sample t-tests
}
\author{
Angela Andreella
}
