% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNullDistribution.R
\name{plotNullDistribution}
\alias{plotNullDistribution}
\title{Plot Permutation Pvalues}
\usage{
plotNullDistribution(P,family,alpha, ct, path, name, delta,copes,mask)
}
\arguments{
\item{P}{permutation matrix pvalues where rows correspond to the permutation, default NULL}

\item{family}{which family for the confidence envelope? simes, finner, beta or higher.criticism. default is NULL}

\item{alpha}{type I error allowed to construct the critical family, default 0.1}

\item{ct}{set of threshold to construct the critical family, default c(0,1)}

\item{path}{selected the path where the plot will be saved, if NULL the current working directory is used}

\item{name}{name plot pdf, if NULL plot is used}

\item{delta}{for the family critical values, default NULL}

\item{copes}{image copes instead of pvalues, default NULL}

\item{mask}{mask}

\item{alternative}{character referring to the alternative hypothesis, "two.sided", "greater" or "lower". Default is "two.sided"}

\item{rand}{logical. Should p values computed by permutation distribution?}

\item{B}{number of permutations to perform, default is 1000, if matrix p-value is computed directly.}
}
\value{
Returns plot null distribution with critical value curve and observed pvalues in red
}
\description{
create plot permutation pvalues with corresponding critical vectors
}
