% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNullDistribution.R
\name{plotNullDistribution}
\alias{plotNullDistribution}
\title{Plot Permutation Pvalues}
\usage{
plotNullDistribution(P,family,alpha, path, name, delta,copes,mask, alternative, rand, B)
}
\arguments{
\item{P}{permutation matrix pvalues where rows correspond to the variables, default NULL}

\item{family}{by default \code{family="simes"}. Choose a family of confidence envelopes to compute the critical vector from \code{"simes"}, \code{"aorc"}, \code{"beta"} and \code{"higher.criticism"}.}

\item{alpha}{alpha level}

\item{path}{path used to save the NIfTI file, the path does not must end with \code{/}.}

\item{name}{choose the name of the NIfTI file}

\item{delta}{for the family critical values, default 0. Vectors accepted}

\item{copes}{image copes instead of pvalues, default NULL}

\item{mask}{3D array of locicals (i.e. \code{TRUE/FALSE} in/out of the brain). 
Alternatively it may be a (character) NIfTI file name. If \code{mask=NULL}, it is assumed that non of the voxels have to be excluded.}

\item{alternative}{a character string referring to the alternative hypothesis, must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"lower"}.}

\item{rand}{by default \code{rand = FALSE}.}

\item{B}{by default \code{B = 1000}. Number of permutations.}
}
\value{
Returns plot null distribution with critical value curve and observed pvalues in red
}
\description{
create plot permutation pvalues with corresponding critical vectors
}
