% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowVariance.R
\name{rowVariance}
\alias{rowVariance}
\title{Rows Variance}
\usage{
rowVariance(X,na.rm = TRUE)
}
\arguments{
\item{X}{data matrix where rows represent the \eqn{m} variables and columns the \eqn{n} observations.}

\item{na.rm}{Boolean value. Default @TRUE.}
}
\value{
vector with length equals \eqn{m} rows variance.
}
\description{
performs the variance for each row in a matrix.
}
\author{
Angela Andreella
}
