% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowVariance.R
\name{rowVariance}
\alias{rowVariance}
\title{Rows Variance}
\usage{
rowVariance(X, na.rm = TRUE)
}
\arguments{
\item{X}{data where rows represents the variables and columns the observations}

\item{na.rm}{remove na?}
}
\value{
rows variance
}
\description{
Performs the variance for each row in a matrix
}
\author{
Angela Andreella
}
