% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{simulate normal distributed data}
\usage{
simulateData(pi0,m,n,rho, set.seed = NULL, power, alpha)
}
\arguments{
\item{pi0}{proportion of true null hypothesis}

\item{m}{number of variables, i.e., hypothesis}

\item{n}{number of observations}

\item{rho}{Level of equi-correlation between pairs of variables}

\item{set.seed}{specify seed}

\item{power}{power}

\item{alpha}{alpha value for power analysis}
}
\value{
Returns the simulated data matrix.
}
\description{
Simulate normal distributed data
}
\author{
Angela Andreella
}
