% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{simulate normal distributed data}
\usage{
simulateData(pi0,m,n, rho, seed = NULL, power = 0.8, alpha = 0.05)
}
\arguments{
\item{pi0}{numeric value in `[0,1]`. Proportion of true null hypothesis.}

\item{m}{numeric value. Number of variables.}

\item{n}{numeric value. Number of observations.}

\item{rho}{numeric value in `[0,1]`. Level of equi-correlation between pairs of variables.}

\item{seed}{integer value. If you want to specify the seed. Default to @NULL}

\item{power}{numeric value in `[0,1]`. Level of power. Default 0.8.}

\item{alpha}{numeric value in `[0,1]`. It expresses the alpha level to control the family-wise error rate. Default 0.05.}
}
\value{
Returns a matrix with dimensions \eqn{m \times n}.
}
\description{
Simulate normal distributed data.
}
\author{
Angela Andreella
}
