% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{simulate normal distributed data}
\usage{
simulateData(pi0,m,n, rho, seed = NULL, power = 0.8, alpha = 0.05)
}
\arguments{
\item{pi0}{Numeric value in `[0,1]`. Proportion of true null hypothesis.}

\item{m}{Numeric value. Number of variables.}

\item{n}{Numeric value. Number of observations.}

\item{rho}{Numeric value in `[0,1]`. Level of equi-correlation between pairs of variables.}

\item{seed}{Integer value. If you want to specify the seed. Default to to \code{NULL}}

\item{power}{Numeric value in `[0,1]`. Level of power. Default to 0.8.}

\item{alpha}{Numeric value in `[0,1]`. \eqn{\alpha} level to control the family-wise error rate. Default to 0.05.}
}
\value{
Returns a matrix with dimensions \eqn{m \times n}.
}
\description{
Simulate normal distributed data.
}
\author{
Angela Andreella
}
