% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statmap.R
\name{Statmap}
\alias{Statmap}
\title{Statmap}
\usage{
Statmap(copes,alternative,path,name, Pmap, mask)
}
\arguments{
\item{copes}{list of copes}

\item{alternative}{which alternative hypothesis perform, default is two.sided}

\item{path}{path to save the nifti file, the path doesn' t must end with /}

\item{name}{choose the name of your nifti file}

\item{Pmap}{if you want also the Pvalue map insert TRUE, default is FALSE}

\item{mask}{mask map}
}
\value{
lambda
}
\description{
create Statmap nifti image
}
\author{
Angela Andreella
}
